+++ aeronet/objects/include/aeronet/header-write.hpp
@@ -3,69 +3,66 @@
 #include <cassert>
 #include <charconv>
 #include <concepts>
-#include <cstring>
+#include <limits>
 #include <string_view>
 
+#ifndef NDEBUG
+#include <system_error>
+#endif
+
 #include "aeronet/http-constants.hpp"
-#include "aeronet/nchars.hpp"
+#include "aeronet/memory-utils.hpp"
 #include "aeronet/timedef.hpp"
 #include "aeronet/timestring.hpp"
 
 namespace aeronet {
 
 constexpr char *WriteHeader(char *insertPtr, std::string_view key, std::string_view value) {
-  std::memcpy(insertPtr, key.data(), key.size());
-  std::memcpy(insertPtr + key.size(), http::HeaderSep.data(), http::HeaderSep.size());
+  insertPtr = Append(key, insertPtr);
+  insertPtr = Append(http::HeaderSep, insertPtr);
   if (!value.empty()) {
-    std::memcpy(insertPtr + key.size() + http::HeaderSep.size(), value.data(), value.size());
+    insertPtr = Append(value, insertPtr);
   }
-  return insertPtr + key.size() + http::HeaderSep.size() + value.size();
+  return insertPtr;
 }
 
 constexpr char *WriteHeader(char *insertPtr, std::string_view key, std::integral auto value) {
-  std::memcpy(insertPtr, key.data(), key.size());
-  std::memcpy(insertPtr + key.size(), http::HeaderSep.data(), http::HeaderSep.size());
-  const auto valueSz = nchars(value);
-  insertPtr += key.size() + http::HeaderSep.size();
-  [[maybe_unused]] const auto result = std::to_chars(insertPtr, insertPtr + valueSz, value);
-  assert(result.ec == std::errc() && result.ptr == insertPtr + valueSz);
-  return insertPtr + valueSz;
-}
-
-constexpr char *WriteCRLF(char *insertPtr) {
-  std::memcpy(insertPtr, http::CRLF.data(), http::CRLF.size());
-  return insertPtr + http::CRLF.size();
+  insertPtr = Append(key, insertPtr);
+  insertPtr = Append(http::HeaderSep, insertPtr);
+  const auto [ptr, ec] =
+      std::to_chars(insertPtr, insertPtr + std::numeric_limits<decltype(value)>::digits10 + 1, value);
+  assert(ec == std::errc());
+  return ptr;
 }
 
 // Write an HTTP header field to the given buffer, including a last CRLF.
 // Returns the pointer immediately after the last written byte.
 // Header key must not be empty, but header value may be empty.
 constexpr char *WriteHeaderCRLF(char *insertPtr, std::string_view key, std::string_view value) {
-  return WriteCRLF(WriteHeader(insertPtr, key, value));
+  return Append(http::CRLF, WriteHeader(insertPtr, key, value));
 }
 
 constexpr char *WriteHeaderCRLF(char *insertPtr, std::string_view key, std::integral auto value) {
-  return WriteCRLF(WriteHeader(insertPtr, key, value));
+  return Append(http::CRLF, WriteHeader(insertPtr, key, value));
 }
 
 // Same as above, but CRLF is first
 constexpr char *WriteCRLFHeader(char *insertPtr, std::string_view key, std::string_view value) {
-  return WriteHeader(WriteCRLF(insertPtr), key, value);
+  return WriteHeader(Append(http::CRLF, insertPtr), key, value);
 }
 
 constexpr char *WriteCRLFHeader(char *insertPtr, std::string_view key, std::integral auto value) {
-  return WriteHeader(WriteCRLF(insertPtr), key, value);
+  return WriteHeader(Append(http::CRLF, insertPtr), key, value);
 }
 
 // Write a Date HTTP header field to the given buffer, including a last CRLF.
 // Returns the pointer immediately after the last written byte.
-// Given buffer requires a size of at least "Date".size() + HeaderSep.size() + kRFC7231DateStrLen + CRLF.size().
+// Given buffer requires a size of at least "Date".size() + HeaderSep.size() + RFC7231DateStrLen + CRLF.size().
 constexpr char *WriteCRLFDateHeader(char *insertPtr, SysTimePoint tp) {
-  insertPtr = WriteCRLF(insertPtr);
-  std::memcpy(insertPtr, http::Date.data(), http::Date.size());
-  std::memcpy(insertPtr + http::Date.size(), http::HeaderSep.data(), http::HeaderSep.size());
-
-  return TimeToStringRFC7231(tp, insertPtr + http::Date.size() + http::HeaderSep.size());
+  insertPtr = Append(http::CRLF, insertPtr);
+  insertPtr = Append(http::Date, insertPtr);
+  insertPtr = Append(http::HeaderSep, insertPtr);
+  return TimeToStringRFC7231(tp, insertPtr);
 }
 
 }  // namespace aeronet